% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualFunction.R
\name{visual_function}
\alias{visual_function}
\title{Text Visualizer Function}
\usage{
visual_function(mydata)
}
\arguments{
\item{mydata}{A string input or string vector that is used to find the most common words within.}
}
\description{
This function allows you to visualize the frequency of words within your input text.
}
\examples{
visual_function()
}
\keyword{visualization}
